Remove unused code; Rearrange Q.remove()
There was some unused code in removeExisting() that was confusing
(particularly the bit w/ the job.user and job.raffle_num :D). I removed
it b/c it wasn't needed at all.

I also moved the unmapping of the ID into removeExisting, since that
function's job is "remove a job w/ the custom title XYZ". Unmapping XYZ
is part of that job.